dataplex-extract-data-quality-rules/requirements.txt
google-cloud-dataplex==2.15.0
PyYAML==6.0.3
